pyproject.toml
[tool.poetry]
name = "ftmlabdaq"
version = "0.1.0"
description = ""
authors = ["Antonello Pellecchia <antonello.pellecchia@cern.ch>"]

[[tool.poetry.packages]]
include = "ftm"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
pyserial = "^3.5"
PyYAML = "^6.0"
mplhep = "^0.3.23"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
uproot = "^4.2.2"
requests = "^2.27.1"
python-vxi11 = "^0.9"
tqdm = "^4.64.0"

[tool.poetry.scripts]
ftm = "ftm.app:main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
